# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lorekind editorial API",
    "version": "0.1.0-experimental",
    "description": "Versioned resource API over the local evaluation application. One entry and one active proposal per workspace; prior applied proposals retained. Injected authentication. No GitLab delivery, production identity provider or remote recovery is claimed. Mutations return completed results synchronously; pending/failed remote operations are not implemented. Query parameters and pagination are not supported. PATCH replaces content rather than applying JSON Patch. Version 1 route name does not yet imply stable compatibility."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Verified by the deployment adapter. No identity in query/body."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "details": { "type": "array", "items": { "type": "string" } },
              "message": { "type": "string" },
              "issues": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ValidationIssue" }
              }
            },
            "required": ["code", "details"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false
      },
      "ValidationIssue": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "path": {
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "anyOf": [{ "type": "string" }, { "type": "number" }, { "type": "boolean" }]
            }
          },
          "message": { "type": "string" }
        },
        "required": ["code", "path", "params", "message"],
        "additionalProperties": false
      },
...
      "Receipt": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "revision": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 },
          "state": { "type": "string" },
          "proposalId": { "type": "string" }
        },
        "required": ["key", "revision", "state"],
        "additionalProperties": false
      },
...
      "Operation": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string", "enum": ["completed"] },
          "result": { "$ref": "#/components/schemas/Receipt" }
        },
        "required": ["id", "status", "result"],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/operations/{operationId}": {
      "get": {
        "summary": "Recover a persisted receipt; only its initiating principal can read it",
        "security": [{ "bearerAuth": [] }],
        "operationId": "get_operations_operationId",
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 1, "pattern": "^[^/\\\\]+$" },
            "required": true,
            "name": "operationId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Presentation language: en or es (regional tags accepted), weighted by q. English fallback. Codes, paths, content and revisions are never translated."
            },
            "required": false,
            "description": "Presentation language: en or es (regional tags accepted), weighted by q. English fallback. Codes, paths, content and revisions are never translated.",
            "name": "accept-language",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Recover a persisted receipt; only its initiating principal can read it",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Operation" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "400": {
            "description": "Invalid request, JSON, path or unsupported query",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "401": {
            "description": "Unverified identity",
            "headers": {
              "WWW-Authenticate": { "schema": { "type": "string" }, "required": true },
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Current capability denied",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "404": {
            "description": "Resource missing or inaccessible",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "headers": {
              "Allow": { "schema": { "type": "string" }, "required": true },
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "409": {
            "description": "Workflow, idempotency or historical diff conflict",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "412": {
            "description": "Stale revision precondition",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "413": {
            "description": "Request body exceeds 131072 bytes",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "415": {
            "description": "application/json is required",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "422": {
            "description": "Content violates the trusted profile schema",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "428": {
            "description": "If-Match is required",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          },
          "503": {
            "description": "Local storage is busy",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            },
            "headers": {
              "Content-Language": {
                "description": "Resolved presentation locale. Host-level rejections may omit localization headers.",
                "schema": { "type": "string", "enum": ["en", "es"] }
              },
              "Vary": {
                "description": "Responses from the portable handler vary by Accept-Language.",
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
...
  },
...
}